requirements.txt
# requirements.txt
anyio==4.3.0
APScheduler==3.10.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==5.3.0
dnspython==2.6.1
Flask==2.3.3
geographiclib==2.0
geopy==2.4.1
gevent==23.9.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3  
httpx==0.24.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
pymongo==4.6.2
python-telegram-bot==20.5
pytz==2024.1
redis==5.0.1
requests==2.32.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==2.3.7